+++ packages/mds-api-helpers/index.ts
@@ -144,58 +144,3 @@ export const asJsonApiLinks = (req: express.Request, skip: number, take: number,
   }
   return undefined
 }
-
-export class ServerError extends Error {
-  public constructor(message: string = 'server_error', public info = {}) {
-    super(message)
-    this.name = 'ServerError'
-    this.info = info
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, ServerError)
-    }
-  }
-}
-
-export class ValidationError extends Error {
-  public constructor(message: string = 'validation_error', public info = {}) {
-    super(message)
-    this.name = 'ValidationError'
-    this.info = info
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, ValidationError)
-    }
-  }
-}
-
-export class NotFoundError extends Error {
-  public constructor(message: string = 'not_found_error', public info = {}) {
-    super(message)
-    this.name = 'NotFoundError'
-    this.info = info
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, NotFoundError)
-    }
-  }
-}
-
-export class ConflictError extends Error {
-  public constructor(message: string = 'conflict_error', public info = {}) {
-    super(message)
-    this.name = 'ConflictError'
-    this.info = info
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, ConflictError)
-    }
-  }
-}
-
-export class AuthorizationError extends Error {
-  public constructor(message: string = 'authorization_error', public info = {}) {
-    super(message)
-    this.name = 'AuthorizationError'
-    this.info = info
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, AuthorizationError)
-    }
-  }
-}
