+++ packages/mds-audit/package.json
@@ -21,7 +21,6 @@
   ],
   "author": "City of Los Angeles",
   "dependencies": {
-    "@hapi/joi": "15.1.0",
     "express": "4.17.1",
     "mds-api-authorizer": "0.1.3",
     "mds-api-helpers": "0.1.3",
