+++ packages/mds-compliance/validators.ts
@@ -3,7 +3,7 @@ import * as Joi from '@hapi/joi'
 import { VEHICLE_EVENTS, VEHICLE_TYPES } from 'mds-enums'
 
 import { Policy, Geography, VehicleEvent } from 'mds'
-import { ValidationError } from './exceptions'
+import { ValidationError } from 'mds-utils'
 
 const DAYS_OF_WEEK = ['sun', 'mon', 'tue', 'wed', 'thu', 'fri', 'sat']
