+++ packages/mds-policy/api.ts
@@ -21,8 +21,7 @@ import joiToJsonSchema from 'joi-to-json-schema'
 import { Policy, UUID, Geography } from 'mds'
 import db from 'mds-db'
 import { VEHICLE_TYPES } from 'mds-enums'
-import { isUUID, now, pathsFor } from 'mds-utils'
-import { ServerError } from 'mds-api-helpers'
+import { isUUID, now, pathsFor, ServerError } from 'mds-utils'
 import log from 'mds-logger'
 import { PolicyApiRequest, PolicyApiResponse } from './types'
