+++ packages/mds-utils/package.json
@@ -10,6 +10,7 @@
   "author": "City of Los Angeles",
   "license": "Apache-2.0",
   "dependencies": {
+    "@hapi/joi": "15.1.0",
     "circle-to-polygon": "1.0.2",
     "point-in-polygon": "1.0.1"
   },
@@ -20,6 +21,6 @@
     "test": "yarn test:prettier && yarn test:eslint && yarn test:unit",
     "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
     "test:prettier": "prettier --write --check --ignore-path ../../.gitignore '**/*.ts'",
-    "test:unit": "nyc --check-coverage --exclude tests --extension .ts --lines 5 --reporter=text --reporter=html ts-mocha --require source-map-support/register --recursive tests/**/*.ts"
+    "test:unit": "nyc --check-coverage --exclude tests --extension .ts --lines 30 --reporter=text --reporter=html ts-mocha --require source-map-support/register --recursive tests/**/*.ts"
   }
 }
